feat: add c binding for LDLogLevel_Name
