+++ libs/common/src/CMakeLists.txt
@@ -52,6 +52,7 @@ add_library(${LIBNAME} OBJECT
         bindings/c/flag_listener.cpp
         bindings/c/memory_routines.cpp
         bindings/c/data_source/error_info.cpp
+        bindings/c/logging/log_level.cpp
         log_level.cpp
         config/persistence_builder.cpp
         config/logging_builder.cpp
